+++ CHANGELOG.md
@@ -7,7 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### `Added`
 
-- Added MultiQC box plot of RT alignment residuals in the quant workflow [#462](https://github.com/nf-core/mhcquant/pull/462)
+- Added MultiQC box plot of aligned residuals in the quant workflow [#462](https://github.com/nf-core/mhcquant/pull/462)
 
 ### `Fixed`